remove roiextractors dependency from video interface
