+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 * Add metadata support for `DeepLabCutInterface` [PR #1319](https://github.com/catalystneuro/neuroconv/pull/1319)
 * Added a RecordingInterface for WhiteMatter ephys data [PR #1297](https://github.com/catalystneuro/neuroconv/pull/1297) [PR #1333](https://github.com/catalystneuro/neuroconv/pull/1333)
 * Improved `ScanImageInteface` to support both single and multi-file data [PR #1330](https://github.com/catalystneuro/neuroconv/pull/1330)
-* Enable chunking for `InternalVideoInterfacd` [PR #TBD](https://github.com/catalystneuro/neuroconv/pull/TBD)
+* Enable chunking for `InternalVideoInterfac` [PR #1338](https://github.com/catalystneuro/neuroconv/pull/1338)
 
 
 ## Improvements
